+++ crates/nostr/src/event/mod.rs
@@ -11,6 +11,7 @@ use core::cmp::Ordering;
 use core::fmt;
 use core::hash::{Hash, Hasher};
 
+use secp256k1::constants::SCHNORR_SIGNATURE_SIZE;
 use secp256k1::schnorr::Signature;
 use secp256k1::{Message, Secp256k1, Verification};
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
@@ -300,6 +301,14 @@ impl TryFrom<&Event> for Metadata {
     }
 }
 
+#[inline]
+fn serialize_sig<S: Serializer>(sig: &Signature, s: S) -> Result<S::Ok, S::Error> {
+    let bytes: [u8; SCHNORR_SIGNATURE_SIZE] = sig.serialize();
+    let mut hex_buf = [0u8; SCHNORR_SIGNATURE_SIZE * 2];
+    let hex_str = faster_hex::hex_encode(&bytes, &mut hex_buf).expect("Buffer size is correct");
+    s.serialize_str(hex_str)
+}
+
 /// Struct used for de/serialization of [`Event`]
 #[derive(Serialize, Deserialize)]
 struct EventIntermediate<'a> {
@@ -309,6 +318,7 @@ struct EventIntermediate<'a> {
     pub kind: Cow<'a, Kind>,
     pub tags: Cow<'a, Tags>,
     pub content: Cow<'a, str>,
+    #[serde(serialize_with = "serialize_sig")]
     pub sig: Cow<'a, Signature>,
 }